+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Display "Automated" badge if account is a bot.
 
+### Changed
+
+- Changed arrow shape on landing page.
+
 ## [2.9.0] - 2024-02-07
 
 ### Added